// src/Indice.AspNetCore.Identity.AdminUI.App/src/app/core/models/roles.ts
export const enum RoleNames {
    Administrator = 'Administrator',
    AdminUIAdministrator = 'AdminUIAdministrator',
    AdminUIUsersReader = 'AdminUIUsersReader',
    AdminUIUsersWriter = 'AdminUIUsersWriter',
    AdminUIClientsReader = 'AdminUIClientsReader',
    AdminUIClientsWriter = 'AdminUIClientsWriter'
}